New scripts and massive cleanup
This is the initial import of a script based on arti which also
simplifies things massively considering the current process to be
complicated and time consuming.

Relay operator won't opt-in or opt-out anymore. We now select 200 random
relays matching these criteria:

	- Fast
	- Stable
	- DirCache set to 1
	- Last seen was at least 90 days

The way to opt-out of this is to set your DirCache to 0. We can maybe
one day have mechanism to manually opt-out relays but for now, we don't
offer this option

Signed-off-by: David Goulet <dgoulet@torproject.org>